Skip past unterminated =? escape string
This fixes a panic where we stupidly advance our byte index by 2
when encountering an unterminated escape. Doing this is inefficient
because we'll probably just find the same unterminated escape a bunch
of times, but can also put our byte index in the middle of a UTF-8
character which causes a panic.

Fixes #41