Improve raw header ergonomics slightly
Fixes #90.

When parsing trivial mails (where the headers are
known to be 7-bit, single-line values), it's convenient
if mailparse strips out the trailing '\r' and excludes
it from the raw header value. This patch accomplishes
that. Per spec the line terminator is \r\n so the "raw"
value (which is an implementation detail anyway) could
arguably be said to exclude the \r.