Only match boundaries at the start of a line
Per RFC 2046, the boundary must be at the start of a line
in order to be considered valid. If the boundary exists
somewhere inside the enclosing multipart stuff (which is a
spec violation, but apparently occurs in the wild) the code
would behave unexpectedly. This patch makes it behave better.

Fixes #100